fix: regressions in dc_msg_get_duration() and dc_delete_contact() (#73)
* fix: regression in dc_msg_get_duration, should return int

* fix: regression in dc_delete_contact()

+++ deltachat-ffi/src/lib.rs
@@ -1290,7 +1290,7 @@ pub unsafe extern "C" fn dc_msg_get_height(msg: *mut dc_msg::dc_msg_t) -> libc::
 
 #[no_mangle]
 pub unsafe extern "C" fn dc_msg_get_duration(msg: *mut dc_msg::dc_msg_t) -> libc::c_int {
-    dc_msg::dc_msg_get_duration(msg) as libc::c_int
+    dc_msg::dc_msg_get_duration(msg)
 }
 
 #[no_mangle]
